Move `dispatch.h` to `oif/internal/dispatch.h` and `dispatch_api.h` to `internal/bridge_api.h`
I came with terminology Bridge for components that were called previously
"Language-specific dispatches", hence, it would be better when the API
that they implement is called `bridge_api.h`: it is less confusing.
