+++ oif_impl/lang_c/dispatch_c.c
@@ -6,7 +6,7 @@
 #include <string.h>
 
 #include <oif/api.h>
-#include <oif/dispatch_api.h>
+#include <oif/internal/bridge_api.h>
 #include <oif/allocation_tracker.h>
 #include <oif/config_dict.h>
 #include <oif/util.h>
